// client/src/shared/ui/Button/Button.enums.ts
export enum ButtonColor {
    BLUE = 'blue',
    GREY = 'grey',
    DARK_BLUE = 'dark-blue',
    WHITE = 'white',
    RED = 'red'
}

export enum ButtonIcon {
    SIGN_UP_ARROW = 'SignUpArrow',
    GOOGLE = 'Google',
    FACEBOOK = 'Facebook'
}

export enum ButtonType {
    SMALL = 'small',
    MEDIUM = 'medium',
    MEDIUM_FLEX = 'medium-flex',
    LARGE = 'large',
    LARGE_FLEX = 'large-flex',
    NAV_LOGIN = 'nav-login',
    ROUNDED_SMALL = 'rounded-small',
    SEARCH = 'search'
}